requirements.txt
PyPDF2
python-docx
openai
python-dotenv
